refactor: launch.go 관심사별 분할 및 심링크 로직 통합
- 1,053줄 launch.go 를 4개 파일로 분리 (패키지·함수 시그니처 변경 없음):
  - launch_claudemd.go: buildRootCLAUDEMD, discoverAgentsByDomain, formatProjectsJSON
  - launch_commands.go: 슬래시 커맨드 생성·매니페스트 관리 (commands/*.md embed 포함)
  - launch_symlinks.go: 에이전트 생성·스킬 주입·심링크 관리
  - launch_hooks.go: settings.json hooks 병합 (hooks.json embed 포함)
- launch.go 는 세션 실행 오케스트레이션만 남겨 276줄로 축소
- 중복이던 심링크 동기화 2벌을 통합: syncClaudeAgentLinks 가
  ensureSymlink 를 사용하도록 변경하고 launch_symlinks.go 로 이동
  (심링크 제거 후 재생성, 일반 파일 보존 의미는 동일)
